requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
ollama
sentence-transformers
faiss-cpu
requests

